Conditional render fix, back button half implemented
