+++ felicity/Frontend/src/views/General/Videocall/Video/index.js
@@ -1,9 +1,5 @@
 import React, { useState, useEffect } from "react";
 import { MainContainer, MainVideoContainer, VideoContainer, TextArea, Container, Block, Button, Patient, Name, SubtitleContainer, Record, RecordBox, Group } from "./styles";
-import { IoMdVideocam } from "react-icons/io";
-import { FaPhoneAlt, FaWindows } from "react-icons/fa";
-import { MdKeyboardVoice } from "react-icons/md";
-import { BsFillChatSquareFill } from "react-icons/bs";
 import { Prompt } from 'react-router'
 
 
@@ -14,40 +10,9 @@ const Video = ({ context }) => {
     const [mySpeech, setMySpeech] = useState(false);
     const roleA = JSON.parse(sessionStorage.getItem("role"));
     const name = JSON.parse(sessionStorage.getItem("name"));
-    const [isNavigatingAway, setIsNavigatingAway] = useState(false);
-
-useEffect(() => {
-    const handleBeforeUnload = (event) => {
-    if (!isNavigatingAway) {
-        event.preventDefault();
-        event.returnValue = "";
-    }
-    };
-
-    const handlePopState = () => {
-        setIsNavigatingAway(true);
-    };
-
-    window.addEventListener("beforeunload", handleBeforeUnload);
-    window.addEventListener("popstate", handlePopState);
-
-    return () => {
-        window.removeEventListener("beforeunload", handleBeforeUnload);
-        window.removeEventListener("popstate", handlePopState);
-    };
-}, [isNavigatingAway]);
-
-const handleLeavePage = (location) => {
-    if (!isNavigatingAway) {
-        return "Are you sure you want to leave this page?";
-    }
-};
 
     return (
-        <>
-        <Prompt when={true} message={handleLeavePage} />
-        
-        
+        <>        
         <MainContainer>
 
             {callAccepted && !callEnded && (
